requirements.txt
flask
gunicorn
pandas
numpy
flask-cors
pyarrow
psycopg2-binary  
python-dotenv   
